Modes seem to be complete now. I have added an FSA implementation so that commands can be chained together, such as k3c (Kill 3 clients). The operators are now defined in the operators array in the config file. Their functions begin with op_*.
Motions are also defined in the config header. A motion needs to have a type, right now only CLIENT and WORKSPACE are supported.

The implemenation of operators, counts and motions works for cases such as kc (Kill one client) and k3c (Kill 3 clients). I need to add in error checking and allow the user to exit a state, such as pressing ESC in order to go back to the operator state.