Basic implementation of a stack so that howm can support cutting and pasting of clients.
Made some style tweaks.

Implemented stack_push and stack_pop.

Made the use of postfix and prefix operators clearer.

Added doxygen comments for stack_push and stack_pop.

I need to implement a function to take a segment of a linked list. I also need to unmap windows that are to be stored in the delete buffer.

Added the ability to cut clients out and have them added to the stack.

Implemented cutting and pasting clients. There are some strange bugs when clients are taken from the middle of the client list.

I have made some tweaks to cutting, I can't seem to reproduce the bugs I had before with windows not retiling.

I need to make a new commit that fixed the issue of the wrong windows being focused after a cut or paste.

Don't segfault when trying to cut on an empty workspace.

Added support for cutting and pasting entire workspaces.

Added docs for op_cut and paste as well as fixing the lack of refocusing mentioned in #19.

I have fixed the bug relating to cutting around the end of a client list.

There is a bug when trying to cut more clients than exist.

I also need to make paste update the focused client correctly.

Paste now focuses the last pasted client.

Fixed problems caused by the rebase and added UNUSED to paste's arg.